+++ profile/admm.py
@@ -68,7 +68,7 @@ psf, data = load_data(
     plot=False,
     gray=gray,
     dtype=dtype,
-    torch=True,
+    use_torch=True,
 )
 
 recon = ADMM(psf, dtype=dtype)
@@ -91,7 +91,7 @@ psf, data = load_data(
     plot=False,
     gray=gray,
     dtype=dtype,
-    torch=True,
+    use_torch=True,
     torch_device="cuda",
 )